Move query to op conversion out of optimizer file
