+++ src/backend/optimizer/Makefile.am
@@ -8,6 +8,7 @@ optimizer_FILES = \
 				backend/optimizer/postgres_shim.cpp \
 				backend/optimizer/query_transformer.cpp \
 				backend/optimizer/optimizer.cpp \
+				backend/optimizer/convert_query_to_op.cpp \
 				backend/optimizer/operator_node.cpp \
 				backend/optimizer/query_operators.cpp \
 				backend/optimizer/query_node_printer.cpp \
